The except-handler block doesn't have a block handler
In ceval.c, except-handler is pushed onto the block stack with a handler of -1.  I think this -1 is a C-ism for False/None/not applicable.  We'll never jump to the handler of an except-handler block; except-handler is always unwound by the unwind_except_handler method.

Except-handler is a weird beast, an "implicit block" that doesn't map to any opcode.  (It has a byte value of 257!) See [this thread](https://mail.python.org/pipermail/python-3000/2008-May/013740.html) for an explanation of what's going on here.  As far as I can tell, Antoine's patch was implemented just as he describes in his initial email.