Fix generator implementation
I've come to believe that the right implementation of the VM has one data stack per frame, not one data stack for the VM.  This is consistent with halting and resuming frames using generators.  The bug appeared to be around nested generators, but in fact it should arise with any generator resuming a frame if the stack has been modified in the meantime.